pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fs-xgb"
version = "0.1.0"
description = "Permutation- and SHAP-based feature selection framework for XGBoost"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "Feature Selector Team"}
]
dependencies = [
    "numpy>=1.24",
    "pandas>=1.5",
    "scikit-learn>=1.3",
    "xgboost>=1.7",
    "shap>=0.43",
    "pyyaml>=6.0",
    "matplotlib>=3.7"
]

[project.optional-dependencies]
dev = [
    "ipykernel",
    "jupyter"
]

[tool.setuptools.packages.find]
where = ["."]
